ENH : Upgrade CrownSegmentationcli  from pytorch to pytorch lightning
