+++ CrownSegmentationcli/__CrownSegmentation/net.py
@@ -1,0 +1,216 @@
+
+
+import torch
+from torch import  nn
+
+
+import torchmetrics
+
+from __CrownSegmentation.utils import PolyDataToTensors, CreateIcosahedron
+
+import monai
+from pytorch3d.renderer import (
+        FoVPerspectiveCameras, look_at_rotation, 
+        RasterizationSettings, MeshRenderer, MeshRasterizer, HardPhongShader, AmbientLights,  TexturesVertex, 
+)
+from pytorch3d.structures import Meshes
+
+import pytorch_lightning as pl
+from time import time
+from tqdm import tqdm
+
+class TimeDistributed(nn.Module):
+    def __init__(self, module):
+        super(TimeDistributed, self).__init__()
+        self.module = module
+ 
+    def forward(self, input_seq):
+        assert len(input_seq.size()) > 2
+ 
+        # reshape input data --> (samples * timesteps, input_size)
+        # squash timesteps
+
+        size = input_seq.size()
+
+        batch_size = size[0]
+        time_steps = size[1]
+
+        size_reshape = [batch_size*time_steps] + list(size[2:])
+        reshaped_input = input_seq.contiguous().view(size_reshape)
+ 
+        output = self.module(reshaped_input)
+        
+        output_size = output.size()
+        output_size = [batch_size, time_steps] + list(output_size[1:])
+        output = output.contiguous().view(output_size)
+
+        return output
+
+class MonaiUNet(pl.LightningModule):
+    def __init__(self, out_channels=3, class_weights=None, image_size=320, radius=1.35, subdivision_level=1, train_sphere_samples=4):
+
+        super(MonaiUNet, self).__init__()        
+        
+        self.save_hyperparameters()        
+
+        self.out_channels = out_channels
+        self.class_weights = None
+        if(class_weights is not None):
+            self.class_weights = torch.tensor(class_weights).to(torch.float32)
+            
+        self.loss = monai.losses.DiceCELoss(include_background=False, to_onehot_y=True, softmax=True, ce_weight=self.class_weights)
+        self.accuracy = torchmetrics.Accuracy(task='multiclass',num_classes=out_channels)
+        
+        unet = monai.networks.nets.UNet(
+            spatial_dims=2,
+            in_channels=4,   # images: torch.cuda.FloatTensor[batch_size,224,224,4]
+            out_channels=out_channels, 
+            channels=(16, 32, 64, 128, 256),
+            strides=(2, 2, 2, 2),
+            num_res_units=2,
+        )
+        self.model = TimeDistributed(unet)
+
+        ico_verts, ico_faces, ico_edges = PolyDataToTensors(CreateIcosahedron(radius=radius, sl=subdivision_level))
+        ico_verts = ico_verts.to(torch.float32)
+
+        for idx, v in enumerate(ico_verts):
+            if (torch.abs(torch.sum(v)) == radius):
+                ico_verts[idx] = v + torch.normal(0.0, 1e-7, (3,))
+
+        
+        self.register_buffer("ico_verts", ico_verts)
+
+        cameras = FoVPerspectiveCameras()
+        raster_settings = RasterizationSettings(
+            image_size=image_size, 
+            blur_radius=0, 
+            faces_per_pixel=1,
+            max_faces_per_bin=200000
+        )        
+        rasterizer = MeshRasterizer(
+            cameras=cameras, 
+            raster_settings=raster_settings
+        )
+        lights = AmbientLights()
+        self.renderer = MeshRenderer(
+                rasterizer=rasterizer,
+                shader=HardPhongShader(cameras=cameras, lights=lights)
+        )
+
+    def configure_optimizers(self):
+        optimizer = torch.optim.AdamW(self.parameters(), lr=1e-4)
+        return optimizer
+
+    def to(self, device=None):
+        self.renderer = self.renderer.to(device)
+        return super().to(device)
+
+    def forward(self, x):
+
+        V, F, CN = x
+        
+        X, PF = self.render(V, F, CN)
+        x = self.model(X)
+        
+        return x, X, PF
+
+    def render(self, V, F, CN):
+
+        textures = TexturesVertex(verts_features=CN)
+        meshes = Meshes(verts=V, faces=F, textures=textures)        
+
+        X = []
+        PF = []
+        start = time()
+
+        for idx , camera_position in enumerate(self.ico_verts):
+
+            camera_position = camera_position.unsqueeze(0)
+
+            R = look_at_rotation(camera_position, device=self.device)  # (1, 3, 3)
+            T = -torch.bmm(R.transpose(1, 2), camera_position[:,:,None])[:, :, 0]   # (1, 3)
+
+            images = self.renderer(meshes_world=meshes.clone(), R=R, T=T)
+        
+            fragments = self.renderer.rasterizer(meshes.clone())
+            pix_to_face = fragments.pix_to_face
+            zbuf = fragments.zbuf
+
+            images = torch.cat([images[:,:,:,0:3], zbuf], dim=-1)
+            images = images.permute(0,3,1,2)
+
+            pix_to_face = pix_to_face.permute(0,3,1,2)
+            
+            X.append(images.unsqueeze(1))
+            PF.append(pix_to_face.unsqueeze(1))
+        
+        X = torch.cat(X, dim=1)
+        PF = torch.cat(PF, dim=1)        
+
+        return X, PF
+
+    def training_step(self, train_batch, batch_idx):
+
+        V, F, YF, CN = train_batch
+
+        V = V.to(self.device, non_blocking=True)
+        F = F.to(self.device, non_blocking=True)
+        YF = YF.to(self.device, non_blocking=True)
+        CN = CN.to(self.device, non_blocking=True).to(torch.float32)
+
+        x, X, PF = self((V, F, CN))
+
+        y = torch.take(YF, PF).to(torch.int64)*(PF >= 0)
+
+        x = x.permute(0, 2, 1, 3, 4) #batch, time, channels, H, W -> batch, channels, time, H, W
+        y = y.permute(0, 2, 1, 3, 4)
+            
+        loss = self.loss(x, y)
+
+        batch_size = V.shape[0]
+        self.log('train_loss', loss, batch_size=batch_size)
+        self.accuracy(torch.argmax(x, dim=1, keepdim=True).reshape(-1, 1), y.reshape(-1, 1).to(torch.int32))
+        self.log("train_acc", self.accuracy, batch_size=batch_size)
+
+        return loss
+
+
+
+    def validation_step(self, val_batch, batch_idx):
+        V, F, YF, CN = val_batch
+
+        V = V.to(self.device, non_blocking=True)
+        F = F.to(self.device, non_blocking=True)
+        YF = YF.to(self.device, non_blocking=True)
+        CN = CN.to(self.device, non_blocking=True).to(torch.float32)
+
+        x, X, PF = self((V, F, CN))
+
+        y = torch.take(YF, PF).to(torch.int64)*(PF >= 0)
+
+        x = x.permute(0, 2, 1, 3, 4) #batch, time, channels, H, W -> batch, channels, time, H, W
+        y = y.permute(0, 2, 1, 3, 4)
+            
+        loss = self.loss(x, y)
+        
+        batch_size = V.shape[0]
+        self.accuracy(torch.argmax(x, dim=1, keepdim=True).reshape(-1, 1), y.reshape(-1, 1).to(torch.int32))
+        self.log("val_acc", self.accuracy, batch_size=batch_size, sync_dist=True)
+        self.log('val_loss', loss, batch_size=batch_size, sync_dist=True)
+
+    def test_step(self, batch, batch_idx):
+
+        V, F, YF, CN = batch
+
+        x, X, PF = self(V, F, CN)
+        y = torch.take(YF, PF).to(torch.int64)*(PF >= 0)
+        
+        x = x.permute(0, 2, 1, 3, 4) #batch, time, channels, h, w -> batch, channels, time, h, w
+        y = y.permute(0, 2, 1, 3, 4) 
+
+        loss = self.loss(x, y)
+
+        self.accuracy(torch.argmax(x, dim=1, keepdim=True).reshape(-1, 1), y.reshape(-1, 1).to(torch.int32))        
+
+        return {'test_loss': loss, 'test_correct': self.accuracy}
